add property ply_types to the composite model. Extract some helper functions from _composites/_composite_model_impl.py and move to a separate file
